+++ src/FolderStructure/FolderStructure.tsx
@@ -1,64 +1,95 @@
 import {
-  FileProps,
+  IndentSize,
   FileType,
   FolderProps,
   FolderStructureProps,
   FolderType,
 } from "./FolderStructure.types";
-import Indentation from "./Indentation/Indentation";
 
 const isFolder = (node: FolderType | FileType): node is FolderType => {
   return "children" in node;
 };
 
-const File = ({ name, level = 0 }: FileProps) => {
-  return (
-    <li>
-      <Indentation level={level} />
-      {name}
-    </li>
-  );
-};
-
 const FolderTitle = ({ name }: { name: string }) => {
-  return (
-    <div className="font-bold bg-slate-300 w-full px-4 py-2 rounded-t-md">
+  return name ? (
+    <div className="font-bold bg-slate-200 dark:bg-slate-900 w-full px-4 py-2 rounded-t-md">
       {name}
     </div>
+  ) : null;
+};
+
+//returns a Folder or a <div> representing a file
+const Node = ({
+  child,
+  level,
+  indentSize,
+}: {
+  child: FolderType | FileType;
+  level: number;
+  indentSize: IndentSize;
+}) => {
+  return isFolder(child) ? (
+    <Folder
+      key={child.name}
+      folder={child}
+      level={level}
+      indentSize={indentSize}
+    />
+  ) : (
+    <div key={child.name}>{child.name}</div>
   );
 };
 
-const Folder = ({ folder, level }: FolderProps) => {
+const Folder = ({ folder, level, indentSize}: FolderProps) => {
+  //a unit of margin is equal to 4px
+  const marginUnit = indentSize * 4;
+
   return (
-    <li>
-      <Indentation level={level} />
-      <span className="font-medium">{folder.name}</span>
-      <ul>
+    <details open style={{ marginLeft: marginUnit * level }}>
+      <summary>
+        <span className="font-medium -translate-x-2">{folder.name}</span>
+      </summary>
+      <div>
         {folder.children.map((child) => {
           return isFolder(child) ? (
-            <Folder folder={child} level={level + 1} />
+            <Folder
+              folder={child}
+              level={level + 1}
+              indentSize={indentSize}
+              key={child.name}
+            />
           ) : (
-            <File name={child.name} level={level + 1} />
+            <div key={child.name} className="ml-4">
+              {child.name}
+            </div>
           );
         })}
-      </ul>
-    </li>
+      </div>
+    </details>
   );
 };
 
-export const FolderStructure = ({ data }: FolderStructureProps) => {
+export const FolderStructure = ({
+  data,
+  indentSize = 2,
+  ...remainingProps
+}: FolderStructureProps) => {
   return (
-    <div className="bg-slate-200 text-slate-950 text-sm mx-auto max-w-2xl my-8 rounded-md">
+    <div
+      className="bg-slate-50 dark:bg-slate-800 text-slate-950 dark:text-slate-200 text-sm mx-auto max-w-2xl my-8 rounded-md border border-slate-300"
+      {...remainingProps}
+    >
       <FolderTitle name={data.name} />
-      <ul className="px-4 py-2">
-        {data.children.map((child) => {
-          return isFolder(child) ? (
-            <Folder folder={child} level={0} />
-          ) : (
-            <File name={child.name} level={0} />
-          );
-        })}
-      </ul>
+      <div className="px-4 py-2 rounded-b-md border-t border-t-slate-300">
+        {data.children?.map((child) => (
+          <Node
+            key={child.name}
+            child={child}
+            level={0}
+            indentSize={indentSize}
+          />
+        ))}
+      </div>
     </div>
   );
 };