+++ src/FolderStructure/Indentation/Indentation.types.ts
@@ -1,5 +1,0 @@
-export interface IndentationProps {
-  spaceAmount?: number;
-  level: number;
-  useDashes?: boolean;
-}